+++ client/src/components/common/NavBar.jsx
