+++ client/src/components/forms/HostGameForm.jsx
@@ -1,6 +1,6 @@
 import { useContext, useState } from "react";
-import {VenuesContext} from '../App'
-import { GamesContext } from "../App";
+import {VenuesContext} from '../../App'
+import { GamesContext } from "../../App";
 import { useNavigate } from 'react-router-dom';
 import NavBar from "./NavBar";
