+++ client/src/components/forms/RegisterForm.jsx
@@ -1,5 +1,5 @@
 import { useContext, useState } from "react"
-import { UsersContext } from "../App";
+import { UsersContext } from "../../App";
 import { useNavigate } from 'react-router-dom';
 
 function RegisterForm () {
